requirements.txt
python-telegram-bot>=20.0
uvicorn>=0.20.0
starlette>=0.25.0
httpx>=0.27.0
pocket-option==0.1.5
